api/pyproject.toml
[project]
name = "service"           # TODO: service name
description = ""
requires-python = ">=3.13"
version = "0.0.1"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.8",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.7.1",
    "google-cloud-logging>=3.11.4",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-propagator-gcp>=1.9.0",
    "opentelemetry-exporter-gcp-trace>=1.9.0",
    "opentelemetry-api>=1.30.0",
    "fastapi-pagination>=0.12.34",
    "psycopg[binary]>=3.2.9",
    "google-adk[a2a]>=1.3.0",
    "greenlet>=3.2.4",
]

[project.scripts]
app = "src:server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "alembic>=1.16.4",
  "coverage>=7.6.10",
  "pre-commit>=4.1.0",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
  "ruff>=0.9.4",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 80

[tool.ruff.lint]
extend-select = ["RUF", "F", "I", "UP", "FAST", "ASYNC"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
